+++ app/ix/core/util/OutputWriter.java
@@ -1,7 +1,0 @@
-package ix.core.util;
-
-/**
- * Created by katzelda on 9/14/16.
- */
-public class OutputWriter {
-}